- Implemented Beacon Logging in order to perform some tests in the (near) future. (I should do something similar to the Desktop Client)

+++ common/src/main/java/pt/unl/fct/di/novalincs/yanux/scavenger/common/logging/IFileLogger.java
@@ -15,7 +15,7 @@ package pt.unl.fct.di.novalincs.yanux.scavenger.common.logging;
 import pt.unl.fct.di.novalincs.yanux.scavenger.common.file.IFileOutput;
 
 public interface IFileLogger extends IFileOutput {
-    void log(int id, IReading IReading);
+    void log(long id, IReading IReading);
 
     void log(IReading IReading);
 }
